requirements.txt
Django>=3.0.6,<3.1
django-cors-headers==3.5.0
djangorestframework==3.12.4
psycopg2>=2.7.5, <2.8.0
drf-yasg
uWSGI>=2.0.18,<2.1
